Add option (`EXTERNAL_NAME`) to `pd_add_external` to set external object name independently of target name
This change introduces an option to specify a custom name for the Pd
external, rather than defaulting to the build target name.

This is useful in cases where the Pd library shares a name with a
dynamic or static library being compiled, which would otherwise cause
conflicts.

For example, the pd-saf project uses the Spatial Audio Framework (SAF),
which compiles a library named saf. In this case, naming the Pd external
saf would cause a naming collision.